添加 commentOrder，inputPosition 开放 keepTheme，loading

+++ src/theme.ts
@@ -1,3 +1,5 @@
+import { pageAttributes as page } from './page-attributes';
+
 export function loadTheme(theme: string, origin: string) {
   return new Promise((resolve) => {
     const link = document.createElement('link');
@@ -8,7 +10,9 @@ export function loadTheme(theme: string, origin: string) {
     document.head.appendChild(link);
 
     addEventListener('message', (event) => {
-      sessionStorage.setItem('beaudar-set-theme', event.data.theme);
+      if (JSON.parse(page.keepTheme)) {
+        sessionStorage.setItem('beaudar-set-theme', event.data.theme);
+      }
       if (event.origin === origin && event.data.type === 'set-theme') {
         link.href = `/stylesheets/themes/${event.data.theme}/beaudar.css`;
       }
